% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_expected_col.R
\name{check_expected_col}
\alias{check_expected_col}
\title{Check that mandatory columns are present in the dataset}
\usage{
check_expected_col(data)
}
\arguments{
\item{data}{Data frame. Raw data about projects.}
}
\value{
An error if some colnames are absent.
}
\description{
Check that mandatory columns are present in the dataset
}
