% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_expected_colnames.R
\name{check_expected_colnames}
\alias{check_expected_colnames}
\title{Check that mandatory columns are present in the dataset}
\usage{
check_expected_colnames(data, lang = c("fr", "de"))
}
\arguments{
\item{data}{Data frame. Raw data about projects.}

\item{lang}{Character. Language, either "fr" or "de".}
}
\value{
An error if some colnames are absent.
}
\description{
Check that mandatory columns are present in the dataset
}
