% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_colnames.R
\name{clean_colnames}
\alias{clean_colnames}
\title{Clean colnames and use more explicit ones for some columns}
\usage{
clean_colnames(data)
}
\arguments{
\item{data}{Data frame. Raw data about projects.}
}
\value{
Data frame with mandatory columns
}
\description{
Clean colnames and use more explicit ones for some columns
}
