% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_disjunctive_col.R
\name{create_disjunctive_col}
\alias{create_disjunctive_col}
\title{Create disjonctive a dijonctive column on the dataset}
\usage{
create_disjunctive_col(data, col, delim = ", ")
}
\arguments{
\item{data}{Data frame. Raw data about projects.}

\item{col}{Symbol. Column to be transformed.}

\item{delim}{Character. Delim of values in the original column.}
}
\value{
Data with the column transformed
}
\description{
Create disjonctive a dijonctive column on the dataset
}
