% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_data_raw_de.R
\docType{data}
\name{data_raw_de}
\alias{data_raw_de}
\title{data_raw_de}
\format{
A data frame with 170 rows and 36 variables:
\describe{
\item{ Steckbrief ID }{  character }
\item{ Titel }{  character }
\item{ Steckbrief-Typ }{  character }
\item{ Beschreibung }{  character }
\item{ Organisation }{  character }
\item{ Kontakt }{  character }
\item{ Ebene }{  character }
\item{ I }{  character }
\item{ II }{  logical }
\item{ III }{  character }
\item{ A }{  character }
\item{ B }{  character }
\item{ C }{  character }
\item{ D }{  character }
\item{ Verlinkte Projekte und Massnahmen }{  numeric }
\item{ Besondere Stärken }{  character }
\item{ Wichtige Erkenntnisse aus der Umsetzung }{  character }
\item{ Weiterführende Informationen }{  character }
\item{ Wirkung }{  character }
\item{ Sprache }{  character }
\item{ Reichweite }{  character }
\item{ Thema }{  character }
\item{ Zielgruppe }{  character }
\item{ Risikofaktoren für Vulnerabilität (nur Pilot Erwachsene) }{  character }
\item{ Besonders vulnerable Gruppen }{  character }
\item{ Multiplikatoren_innen }{  character }
\item{ Lebensübergänge für Lebensphase II (nur Pilot Erwachsene) }{  character }
\item{ Setting }{  character }
\item{ Art der Policy }{  character }
\item{ Art der Vernetzung }{  character }
\item{ Art der Öffentlichkeitsarbeit }{  character }
\item{ Kommunikationskanäle }{  character }
\item{ Intensität der Vernetzung }{  character }
\item{ Partizipation }{  character }
\item{ Verhaltens- und Verhältnisorientierung }{  character }
\item{ Schwerpunkte }{  character }
}
}
\source{
Source
}
\usage{
data_raw_de
}
\description{
Toy raw Deutsch data about projects.
}
\keyword{datasets}
