% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters.R
\name{get_clusters}
\alias{get_clusters}
\title{Analyze data with MCA and HCPC}
\usage{
get_clusters(data)
}
\arguments{
\item{data}{Data frame. Prepared data about projects.}
}
\description{
Analyze data with MCA and HCPC
}
