% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distrib_level_data.R
\name{get_distrib_level_data}
\alias{get_distrib_level_data}
\title{Get the data for the plot about the distribution of levels}
\usage{
get_distrib_level_data(data)
}
\arguments{
\item{data}{Data frame. Prepared data about projects.}
}
\value{
Data frame. Number of observations by level
}
\description{
Get the data for the plot about the distribution of levels
}
