% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_clusters.R
\name{interpret_clusters}
\alias{interpret_clusters}
\title{Interpret the clusters of projects by using a Large Language Model}
\usage{
interpret_clusters(res_analysis)
}
\arguments{
\item{res_analysis}{List. Results of the analysis (MCA + clustering).}
}
\value{
A list with the interpretation of the clusters of projects.
}
\description{
Interpret the clusters of projects by using a Large Language Model
}
