% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_mandatory_col.R
\name{keep_mandatory_col}
\alias{keep_mandatory_col}
\title{Keep only mandatory columns}
\usage{
keep_mandatory_col(data)
}
\arguments{
\item{data}{Data frame. Raw data about projects.}
}
\value{
Data frame with mandatory columns
}
\description{
Keep only mandatory columns
}
