% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_description.R
\name{parse_description}
\alias{parse_description}
\title{Parse the description of the project by using spaCy}
\usage{
parse_description(data, nb_min_word = 10)
}
\arguments{
\item{data}{Data frame. Raw data about projects.}

\item{nb_min_word}{Integer. Minimum number of occurences to keep a word.}
}
\value{
Data frame with parsed description
}
\description{
Parse the description of the project by using spaCy
}
