% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot the clusters on the MCA factorial plan}
\usage{
plot_clusters(
  res_analysis,
  axes = c(1, 2),
  colors = c("#c1e1ec", "#f08c00", "#ced4da", "#fcc2d7", "#99e9f2"),
  title_font_size = 16,
  text_font_size = 7,
  font_family = "DIN Pro",
  language = c("fr", "de"),
  data_raw_de = NULL
)
}
\arguments{
\item{res_analysis}{List. Object obtained from the data analysis.}

\item{axes}{Character. Axes to be plotted.}

\item{colors}{Character. Colors of the clusters.}

\item{title_font_size}{Integer. Font size for the title.}

\item{text_font_size}{Integer. Font size for the texts.}

\item{font_family}{Character.}

\item{language}{Character. "fr" or "de".}

\item{data_raw_de}{Data frame. Raw German data.}
}
\value{
An interactive plot
}
\description{
Plot the clusters on the MCA factorial plan
}
