% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distrib_level.R
\name{plot_distrib_level}
\alias{plot_distrib_level}
\title{Plot the distribution of the levels of projects.}
\usage{
plot_distrib_level(
  data,
  color = "#c1e1ec",
  title_font_size = 16,
  text_font_size = 11,
  font_family = "DIN Pro",
  language = c("fr", "de")
)
}
\arguments{
\item{data}{Data frame. Prepared data about projects.}

\item{color}{Character. Color for the bubbles.}

\item{title_font_size}{Integer. Font size for the title.}

\item{text_font_size}{Integer. Font size for the texts.}

\item{font_family}{Character.}

\item{language}{Character. "fr" or "de".}
}
\value{
An interactive plot
}
\description{
Plot the distribution of the levels of projects.
}
