% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distrib_type.R
\name{plot_distrib_type}
\alias{plot_distrib_type}
\title{Plot the distribution of the types of projects.}
\usage{
plot_distrib_type(
  data,
  color = "#c1e1ec",
  font_family = "Roboto",
  title_font_size = "16px",
  text_font_size = "11px",
  language = c("fr", "de")
)
}
\arguments{
\item{data}{Data frame. Prepared data about projects.}

\item{color}{Character. Color for the bubbles.}

\item{font_family}{Character. Font family.}

\item{title_font_size}{Character. Font size for the title.}

\item{text_font_size}{Character. Font size for the texts.}

\item{language}{Character. "fr" or "de".}
}
\value{
An interactive plot
}
\description{
Plot the distribution of the types of projects.
}
