% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare raw data for statistical data analysis}
\usage{
prepare_data(data)
}
\arguments{
\item{data}{Data frame. Raw data about projects.}
}
\value{
Data frame ready for analysis
}
\description{
Prepare raw data for statistical data analysis
}
