% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_res_interpret_clusters.R
\docType{data}
\name{res_interpret_clusters}
\alias{res_interpret_clusters}
\title{res_interpret_clusters}
\format{
A list
\describe{
\item{ intro_llm }{  character }
\item{ request_llm }{  character }
\item{ res_prompt_llm }{  character }
\item{ res_desc_llm_fr }{  character }
\item{ res_desc_llm_de }{  character }
}
}
\usage{
res_interpret_clusters
}
\description{
Example of results with the interpretation of clusters of projects.
}
\keyword{datasets}
