% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_raw_data.R
\name{show_raw_data}
\alias{show_raw_data}
\title{Show the raw data into an interactive table.}
\usage{
show_raw_data(data, nb_by_page = 2, language = c("fr", "de"))
}
\arguments{
\item{data}{Data frame. Raw data about projects.}

\item{nb_by_page}{Integer. Number of projects to be displayed by page.}

\item{language}{Character. "fr" or "de".}
}
\value{
An interactive table
}
\description{
Show the raw data into an interactive table.
}
