% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distrib_type_data.R
\name{get_distrib_type_data}
\alias{get_distrib_type_data}
\title{Get the data for the plot about the distribution of types}
\usage{
get_distrib_type_data(data)
}
\arguments{
\item{data}{Data frame. Prepared data about projects.}
}
\value{
Data frame. Number of "Oui" by column A/B/C/D
}
\description{
Get the data for the plot about the distribution of types
}
