% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distrib_type.R
\name{plot_distrib_type}
\alias{plot_distrib_type}
\title{Plot the distribution of the types of projects.}
\usage{
plot_distrib_type(
  data,
  color = "#c1e1ec",
  title_font_size = 16,
  text_font_size = 11,
  font_family = "DIN Pro",
  language = c("fr", "de"),
  xmin = 1,
  xmax = 2,
  ymin = 1.3,
  ymax = 2
)
}
\arguments{
\item{data}{Data frame. Prepared data about projects.}

\item{color}{Character. Color for the bubbles.}

\item{title_font_size}{Integer. Font size for the title.}

\item{text_font_size}{Integer. Font size for the texts.}

\item{font_family}{Character.}

\item{language}{Character. "fr" or "de".}

\item{xmin}{Integer.}

\item{xmax}{Integer.}

\item{ymin}{Integer.}

\item{ymax}{Integer.}
}
\value{
An interactive plot
}
\description{
Plot the distribution of the types of projects.
}
